+++ src/components/Layout/Content/BottomContent/BottomContent.jsx
@@ -1,10 +1,8 @@
 import React from 'react'
 import './BottomContent.css'
-import { MOSAIC_MIN_ZOOM, APP_NAME, SearchTypes } from '../../../defaults'
+import { MOSAIC_MIN_ZOOM, APP_NAME } from '../../../defaults'
 import LeafMap from '../../../LeafMap/LeafMap'
-
 import PopupResults from '../../../PopupResults/PopupResults'
-
 import LoadingAnimation from '../../../LoadingAnimation/LoadingAnimation'
 import Legend from '../../../Legend/Legend'
 import {
@@ -12,17 +10,18 @@ import {
   VITE_SHOW_PUBLISH_BTN,
   VITE_ANALYZE_BTN_URL
 } from '../../../../assets/config.js'
-
-// redux imports
 import { useSelector, useDispatch } from 'react-redux'
-// you need to import each action you need to use
 import {
   setShowPublishModal,
   setShowLaunchModal,
-  setShowZoomNotice
+  setShowZoomNotice,
+  setisDrawingEnabled
 } from '../../../../redux/slices/mainSlice'
-
-import { setMapZoomLevel } from '../../../../utils/mapHelper'
+import {
+  setMapZoomLevel,
+  disableMapPolyDrawing
+} from '../../../../utils/mapHelper'
+import Box from '@mui/material/Box'
 
 const BottomContent = () => {
   // set up useSelector to get value from store
@@ -38,6 +37,9 @@ const BottomContent = () => {
   const _searchType = useSelector((state) => state.mainSlice.searchType)
   const _viewMode = useSelector((state) => state.mainSlice.viewMode)
   const _showPopupModal = useSelector((state) => state.mainSlice.showPopupModal)
+  const _isDrawingEnabled = useSelector(
+    (state) => state.mainSlice.isDrawingEnabled
+  )
 
   // if you are setting redux state, call dispatch
   const dispatch = useDispatch()
@@ -71,15 +73,20 @@ const BottomContent = () => {
     }
   }
 
+  function onCancelDrawGeomClicked() {
+    dispatch(setisDrawingEnabled(false))
+    disableMapPolyDrawing()
+  }
+
   return (
     <div className="BottomContent">
       <LeafMap></LeafMap>
       {_showZoomNotice && (
         <div className="ZoomNotice">
           Images are not visible at this zoom level.
-          <a onClick={onZoomClick}>
+          <Box className="zoomInText" onClick={onZoomClick}>
             <strong>Zoom in</strong>
-          </a>
+          </Box>
         </div>
       )}
       <div className="actionButtons">
@@ -100,14 +107,16 @@ const BottomContent = () => {
         )}
       </div>
       {_searchResults?.numberMatched &&
-      _searchResults?.searchType !== 'AggregatedResults' ? (
-        <div className="resultCount">
+      _searchResults?.searchType !== 'AggregatedResults' &&
+      !_isDrawingEnabled ? (
+        <div className="resultCount" data-testid="testShowingScenesMessage">
           Showing {_searchResults.numberReturned} of{' '}
           {_searchResults.numberMatched} scenes
         </div>
       ) : null}
-      {_searchResults?.searchType === 'AggregatedResults' ? (
-        <div className="resultCount">
+      {_searchResults?.searchType === 'AggregatedResults' &&
+      !_isDrawingEnabled ? (
+        <div className="resultCount" data-testid="testShowingAggregatedMessage">
           <strong>Showing Aggregated Results</strong>
           {_searchResults.features.length} {resultType},{' '}
           {_searchResults.numberMatched} total scenes
@@ -133,6 +142,26 @@ const BottomContent = () => {
         _searchResults?.searchType === 'AggregatedResults' && (
           <Legend results={_searchResults}></Legend>
         )}
+      {_isDrawingEnabled ? (
+        <div className="drawGeomMessage">
+          <div className="drawGeomMessageText">
+            <span className="drawGeomMessageTextTitle">
+              Draw Search Boundary
+            </span>
+            <span>
+              Click the map to add points. Click the first point to finish.
+            </span>
+          </div>
+          <div className="drawGeomMessageButtons">
+            <button
+              className="drawGeomMessageCancelButton"
+              onClick={onCancelDrawGeomClicked}
+            >
+              Cancel
+            </button>
+          </div>
+        </div>
+      ) : null}
     </div>
   )
 }