+++ src/components/Search/Search.jsx
@@ -1,21 +1,32 @@
 import { React, useEffect } from 'react'
 import './Search.css'
 import { useDispatch, useSelector } from 'react-redux'
-import { setIsAutoSearchSet } from '../../redux/slices/mainSlice'
+import {
+  setIsAutoSearchSet,
+  setshowAdvancedSearchOptions,
+  setisDrawingEnabled,
+  setsearchGeojsonBoundary
+} from '../../redux/slices/mainSlice'
 import 'react-tooltip/dist/react-tooltip.css'
 import Switch from '@mui/material/Switch'
 import DateTimeRangeSelector from '../DateTimeRangeSelector/DateTimeRangeSelector'
 import CloudSlider from '../CloudSlider/CloudSlider'
 import CollectionDropdown from '../CollectionDropdown/CollectionDropdown'
 import ViewSelector from '../ViewSelector/ViewSelector'
-import { VITE_MOSAIC_TILER_URL } from '../../assets/config'
-
+import {
+  VITE_MOSAIC_TILER_URL,
+  VITE_ADVANCED_SEARCH_ENABLED
+} from '../../assets/config'
+import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
+import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
 import { newSearch, debounceNewSearch } from '../../utils/searchHelper'
+import { Box } from '@mui/material'
+import { enableMapPolyDrawing, clearLayer } from '../../utils/mapHelper'
 
 const Search = () => {
   const dispatch = useDispatch()
   const _selectedCollectionData = useSelector(
-    (state) => state.mainSlice._selectedCollectionData
+    (state) => state.mainSlice.selectedCollectionData
   )
   const _searchDateRangeValue = useSelector(
     (state) => state.mainSlice.searchDateRangeValue
@@ -25,10 +36,19 @@ const Search = () => {
   const _isAutoSearchSet = useSelector(
     (state) => state.mainSlice.isAutoSearchSet
   )
-
+  const _showAdvancedSearchOptions = useSelector(
+    (state) => state.mainSlice.showAdvancedSearchOptions
+  )
+  const _isDrawingEnabled = useSelector(
+    (state) => state.mainSlice.isDrawingEnabled
+  )
+  const _searchGeojsonBoundary = useSelector(
+    (state) => state.mainSlice.searchGeojsonBoundary
+  )
   const mosaicTilerURL = VITE_MOSAIC_TILER_URL || ''
 
   useEffect(() => {
+    dispatch(setshowAdvancedSearchOptions(false))
     if (_isAutoSearchSet) {
       debounceNewSearch()
     }
@@ -43,6 +63,35 @@ const Search = () => {
 
   function processSearchBtn() {
     newSearch()
+    dispatch(setshowAdvancedSearchOptions(false))
+  }
+
+  function onAdvancedOptionsClicked() {
+    dispatch(setshowAdvancedSearchOptions(!_showAdvancedSearchOptions))
+  }
+
+  function onDrawBoundaryClicked() {
+    if (_searchGeojsonBoundary) {
+      return
+    }
+    dispatch(setshowAdvancedSearchOptions(!_showAdvancedSearchOptions))
+    dispatch(setisDrawingEnabled(true))
+    enableMapPolyDrawing()
+  }
+
+  function onClearButtonClicked() {
+    if (!_searchGeojsonBoundary) {
+      return
+    }
+    dispatch(setsearchGeojsonBoundary(null))
+    dispatch(setshowAdvancedSearchOptions(false))
+    clearLayer('drawBoundsLayer')
+  }
+
+  function onAdvancedSearchBlur(event) {
+    if (!event.currentTarget.contains(event.relatedTarget)) {
+      dispatch(setshowAdvancedSearchOptions(false))
+    }
   }
 
   return (
@@ -61,24 +110,107 @@ const Search = () => {
           <ViewSelector></ViewSelector>
         </div>
       )}
-      <div className="searchContainer searchButton">
-        <button
-          className={`actionButton disabled-${_isAutoSearchSet}`}
-          onClick={() => processSearchBtn()}
-          disabled={_isAutoSearchSet}
-        >
-          Search
-        </button>
 
-        <div className="autoSearchContainer">
-          <label>Auto Search</label>
-          <Switch
-            checked={_isAutoSearchSet}
-            onChange={handleSwitchChange}
-            inputProps={{ 'aria-label': 'controlled' }}
-          />
+      {VITE_ADVANCED_SEARCH_ENABLED ? (
+        <Box
+          className={
+            _showAdvancedSearchOptions
+              ? 'searchContainer searchButtonAdvanced ' + 'active'
+              : 'searchContainer searchButtonAdvanced'
+          }
+          onBlur={onAdvancedSearchBlur}
+          tabIndex={0}
+        >
+          <div className="advancedSearchContent">
+            <button
+              className={`actionButton`}
+              onClick={() => processSearchBtn()}
+            >
+              Search
+            </button>
+            <Box
+              className="advancedSearchOptions"
+              onClick={onAdvancedOptionsClicked}
+            >
+              Advanced
+              {_showAdvancedSearchOptions ? (
+                <KeyboardArrowUpIcon></KeyboardArrowUpIcon>
+              ) : (
+                <KeyboardArrowDownIcon></KeyboardArrowDownIcon>
+              )}
+            </Box>
+          </div>
+          {_showAdvancedSearchOptions ? (
+            <Box className="advancedSearchOptionsContainer">
+              <span className="advancedSearchOptionsText">
+                Limit search to boundary
+              </span>
+              <div className="advancedSearchOptionsButtons">
+                <button
+                  className={
+                    !_searchGeojsonBoundary
+                      ? 'advancedSearchOptionsButton'
+                      : 'advancedSearchOptionsButton ' +
+                        'advancedSearchOptionsButtonDisabled'
+                  }
+                  onClick={onDrawBoundaryClicked}
+                >
+                  Draw boundary
+                </button>
+                <button
+                  className={
+                    !_searchGeojsonBoundary
+                      ? 'advancedSearchOptionsButton'
+                      : 'advancedSearchOptionsButton ' +
+                        'advancedSearchOptionsButtonDisabled'
+                  }
+                >
+                  Upload GeoJSON
+                </button>
+                <button
+                  className={
+                    _searchGeojsonBoundary
+                      ? 'advancedSearchOptionsButton'
+                      : 'advancedSearchOptionsButton ' +
+                        'advancedSearchOptionsButtonDisabled'
+                  }
+                  onClick={onClearButtonClicked}
+                >
+                  Clear
+                </button>
+              </div>
+            </Box>
+          ) : null}
+        </Box>
+      ) : (
+        <div className="searchContainer searchButton">
+          <button
+            className={`actionButton disabled-${_isAutoSearchSet}`}
+            onClick={() => processSearchBtn()}
+            disabled={_isAutoSearchSet}
+          >
+            Search
+          </button>
+
+          <div
+            className="autoSearchContainer"
+            data-testid="test_autoSearchContainer"
+          >
+            <label htmlFor="autoSearchSwtich">Auto Search</label>
+            <Switch
+              checked={_isAutoSearchSet}
+              onChange={handleSwitchChange}
+              inputProps={{ 'aria-label': 'controlled' }}
+            />
+          </div>
         </div>
-      </div>
+      )}
+      {_isDrawingEnabled ? (
+        <div
+          className="disableSearchOverlay"
+          data-testid="test_disableSearchOverlay"
+        ></div>
+      ) : null}
     </div>
   )
 }