+++ src/utils/searchHelper.js
@@ -101,18 +101,27 @@ export function newSearch() {
 function buildSearchScenesParams(gridCodeToSearchIn) {
   const _selectedCollection = store.getState().mainSlice.selectedCollectionData
   const bbox = buildUrlParamFromBBOX()
-  const dateTimeRange = convertDateForURL(
+  const _dateTimeRange = convertDateForURL(
     store.getState().mainSlice.searchDateRangeValue
   )
   const limit = VITE_API_MAX_ITEMS || 200
   const collections = _selectedCollection.id
+  const _searchGeojsonBoundary =
+    store.getState().mainSlice.searchGeojsonBoundary
 
   const searchParams = new Map([
-    ['bbox', bbox],
-    ['datetime', dateTimeRange],
+    ['datetime', _dateTimeRange],
     ['limit', limit],
     ['collections', collections]
   ])
+  if (_searchGeojsonBoundary) {
+    searchParams.set(
+      'intersects',
+      encodeURIComponent(JSON.stringify(_searchGeojsonBoundary.geometry))
+    )
+  } else {
+    searchParams.set('bbox', bbox)
+  }
 
   const query = {}
   if (_selectedCollection.queryables['eo:cloud_cover']) {
@@ -159,9 +168,11 @@ function buildSearchScenesParams(gridCodeToSearchIn) {
 function buildSearchAggregateParams(gridType) {
   const _selectedCollection = store.getState().mainSlice.selectedCollectionData
   const bbox = buildUrlParamFromBBOX()
-  const dateTimeRange = convertDateForURL(
+  const _dateTimeRange = convertDateForURL(
     store.getState().mainSlice.searchDateRangeValue
   )
+  const _searchGeojsonBoundary =
+    store.getState().mainSlice.searchGeojsonBoundary
   const collections = _selectedCollection.id
 
   let aggregations
@@ -194,11 +205,18 @@ function buildSearchAggregateParams(gridType) {
   }
 
   const searchParams = new Map([
-    ['bbox', bbox],
-    ['datetime', dateTimeRange],
+    ['datetime', _dateTimeRange],
     ['collections', collections],
     ['aggregations', aggregations]
   ])
+  if (_searchGeojsonBoundary) {
+    searchParams.set(
+      'intersects',
+      encodeURIComponent(JSON.stringify(_searchGeojsonBoundary.geometry))
+    )
+  } else {
+    searchParams.set('bbox', bbox)
+  }
 
   const query = {}
   if (_selectedCollection.queryables['eo:cloud_cover']) {
@@ -400,6 +418,8 @@ function newMosaicSearch() {
   const _selectedCollectionData =
     store.getState().mainSlice.selectedCollectionData
   const datetime = convertDate(store.getState().mainSlice.searchDateRangeValue)
+  const _searchGeojsonBoundary =
+    store.getState().mainSlice.searchGeojsonBoundary
   const bboxFromMap = bboxFromMapBounds()
 
   const createMosaicBody = {
@@ -407,8 +427,12 @@ function newMosaicSearch() {
     asset_name: constructMosaicAssetVal(_selectedCollectionData.id),
     collections: [_selectedCollectionData.id],
     datetime,
-    bbox: bboxFromMap,
     max_items: VITE_MOSAIC_MAX_ITEMS || 100
+  }
+  if (_searchGeojsonBoundary) {
+    createMosaicBody.intersects = _searchGeojsonBoundary.geometry
+  } else {
+    createMosaicBody.bbox = bboxFromMap
   }
 
   if (store.getState().mainSlice.showCloudSlider) {